+++ Extension/src/background/api/filters/custom/main.ts
@@ -215,7 +215,7 @@ export class CustomFilterApi {
         const filterMetadata: CustomFilterMetadata = {
             filterId,
             displayNumber: 0,
-            groupId: AntibannerGroupsId.CustomFilterGroupId,
+            groupId: AntibannerGroupsId.CustomFiltersGroupId,
             name,
             description,
             homepage,