+++ Extension/src/background/services/filters.ts
@@ -16,17 +16,20 @@
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
 import {
-    AddAndEnableFilterMessage,
-    DisableFilterMessage,
-    DisableFiltersGroupMessage,
-    EnableFiltersGroupMessage,
     MessageType,
+    type AddAndEnableFilterMessage,
+    type DisableFilterMessage,
+    type DisableFiltersGroupMessage,
+    type EnableFiltersGroupMessage,
+    type SetConsentedFiltersMessage,
+    type GetIsConsentedFilterMessage,
 } from '../../common/messages';
 import { Log } from '../../common/log';
 import { SettingOption } from '../schema';
 import { messageHandler } from '../message-handler';
 import { Engine } from '../engine';
 import {
+    annoyancesConsent,
     FilterMetadata,
     FiltersApi,
     FilterUpdateApi,
@@ -60,6 +63,8 @@ export class FiltersService {
         messageHandler.addListener(MessageType.DisableFiltersGroup, FiltersService.onGroupDisable);
         messageHandler.addListener(MessageType.CheckFiltersUpdate, FiltersService.manualCheckFiltersUpdate);
         messageHandler.addListener(MessageType.ResetBlockedAdsCount, FiltersService.resetBlockedAdsCount);
+        messageHandler.addListener(MessageType.SetConsentedFilters, FiltersService.setConsentedFilters);
+        messageHandler.addListener(MessageType.GetIsConsentedFilter, FiltersService.getIsConsentedFilter);
 
         contextMenuEvents.addListener(ContextMenuAction.UpdateFilters, FiltersService.manualCheckFiltersUpdate);
 
@@ -202,6 +207,32 @@ export class FiltersService {
      */
     private static async resetBlockedAdsCount(): Promise<void> {
         await PageStatsApi.reset();
+    }
+
+    /**
+     * Called on a request to add filter ids to consented filters list.
+     *
+     * @param message Message of {@link SetConsentedFiltersMessage} with filter ids to add.
+     */
+    private static async setConsentedFilters(message: SetConsentedFiltersMessage): Promise<void> {
+        const { filterIds } = message.data;
+
+        await annoyancesConsent.addFilterIds(filterIds);
+    }
+
+    /**
+     * Called on a request to check if filter is consented.
+     *
+     * @param message Message of {@link GetIsConsentedFilterMessage} with filter id to check.
+     *
+     * @returns True if consent is granted for filter, otherwise false.
+     */
+    private static async getIsConsentedFilter(message: GetIsConsentedFilterMessage): Promise<boolean> {
+        const { filterId } = message.data;
+
+        const isConsented = await annoyancesConsent.isConsentedFilter(filterId);
+
+        return isConsented;
     }
 
     /**