+++ Extension/src/common/messages/constants.ts
@@ -104,6 +104,8 @@ export enum MessageType {
   ShowAlertPopup = 'showAlertPopup',
   ShowVersionUpdatedPopup = 'showVersionUpdatedPopup',
   UpdateListeners = 'updateListeners',
+  SetConsentedFilters = 'setConsentedFilters',
+  GetIsConsentedFilter = 'getIsConsentedFilter',
 }
 
 export type ApplySettingsJsonMessage = {
@@ -383,6 +385,20 @@ export type DisableFiltersGroupMessage = {
   type: MessageType.DisableFiltersGroup,
   data: {
     groupId: number,
+  }
+};
+
+export type SetConsentedFiltersMessage = {
+  type: MessageType.SetConsentedFilters,
+  data: {
+    filterIds: number[],
+  }
+};
+
+export type GetIsConsentedFilterMessage = {
+  type: MessageType.GetIsConsentedFilter,
+  data: {
+    filterId: number,
   }
 };