+++ Extension/src/pages/common/components/ConfirmModal/ConfirmModal.jsx
@@ -19,17 +19,32 @@
 import React from 'react';
 import Modal from 'react-modal';
 
+import cn from 'classnames';
+
+import { translator } from '../../../../common/translators/translator';
+import { Icon } from '../ui/Icon';
+
 export const ConfirmModal = ({
     title,
     subtitle,
     isOpen,
     onConfirm,
+    onCancel,
     setIsOpen,
     customCancelTitle,
     customConfirmTitle,
+    isConsent,
 }) => {
     const confirmTitle = customConfirmTitle || 'OK';
-    const cancelTitle = customCancelTitle || 'Cancel';
+    const cancelTitle = customCancelTitle || translator.getMessage('options_confirm_modal_cancel_button');
+
+    const subtitleClassName = cn('modal__subtitle', {
+        'modal__subtitle--short': !isConsent,
+    });
+
+    const okBtnClassName = cn('button button--m modal__btn button--green', {
+        'button--red-bg': !isConsent,
+    });
 
     const closeModal = () => {
         setIsOpen(false);
@@ -40,35 +55,54 @@ export const ConfirmModal = ({
         onConfirm();
     };
 
+    const handleCancel = () => {
+        closeModal();
+        onCancel();
+    };
+
     return (
         <>
             <Modal
                 isOpen={isOpen}
-                onRequestClose={closeModal}
+                onRequestClose={handleCancel}
+                // reset default padding
+                style={{ content: { padding: 0 } }}
             >
-                <div className="modal__title">
-                    {title}
-                </div>
-                {subtitle && (
-                    <div className="modal__subtitle modal__subtitle--confirm">
-                        {subtitle}
+                <div className="modal">
+                    <div className="modal__header">
+                        <button
+                            type="button"
+                            className="button modal__close"
+                            aria-label={translator.getMessage('close_button_title')}
+                            onClick={handleCancel}
+                        >
+                            <Icon id="#cross" />
+                        </button>
                     </div>
-                )}
-                <div className="modal__content modal__content--button">
-                    <button
-                        className="button button--m button--red-bg modal__btn modal__btn--statistic"
-                        type="button"
-                        onClick={handleConfirm}
-                    >
-                        {confirmTitle}
-                    </button>
-                    <button
-                        className="button button--m button--transparent modal__btn modal__btn--statistic"
-                        type="button"
-                        onClick={closeModal}
-                    >
-                        {cancelTitle}
-                    </button>
+                    <div className="modal__title">
+                        {title}
+                    </div>
+                    {subtitle && (
+                        <div className={subtitleClassName}>
+                            {subtitle}
+                        </div>
+                    )}
+                    <div className="modal__actions">
+                        <button
+                            className={okBtnClassName}
+                            type="button"
+                            onClick={handleConfirm}
+                        >
+                            {confirmTitle}
+                        </button>
+                        <button
+                            className="button button--m button--transparent modal__btn modal__btn--confirm"
+                            type="button"
+                            onClick={handleCancel}
+                        >
+                            {cancelTitle}
+                        </button>
+                    </div>
                 </div>
             </Modal>
         </>