+++ Extension/src/pages/options/stores/SettingsStore.js
@@ -42,7 +42,9 @@ import {
 } from '../components/Filters/helpers';
 import { optionsStorage } from '../options-storage';
 import {
+    AntiBannerFiltersId,
     AntibannerGroupsId,
+    RECOMMENDED_TAG_ID,
     TRUSTED_TAG,
     WASTE_CHARACTERS,
 } from '../../../common/constants';
@@ -118,6 +120,12 @@ class SettingsStore {
 
     @observable allowlistSizeReset = false;
 
+    @observable filtersToGetConsentFor = [];
+
+    @observable isAnnoyancesConsentModalOpen = false;
+
+    @observable filterIdSelectedForConsent = null;
+
     constructor(rootStore) {
         makeObservable(this);
         this.rootStore = rootStore;
@@ -134,6 +142,12 @@ class SettingsStore {
 
     @action
     async requestOptionsData(firstRender) {
+        // do not re-render options page if the annoyances consent modal is open.
+        // it is needed to prevent switch disabling due to the actual state while the modal is shown
+        if (this.isAnnoyancesConsentModalOpen) {
+            return;
+        }
+
         const data = await messenger.getOptionsData();
         runInAction(() => {
             this.settings = data.settings;
@@ -292,6 +306,17 @@ class SettingsStore {
         return category.enabled;
     }
 
+    /**
+     * Checks whether the group is touched.
+     *
+     * @param {number} groupId Group id.
+     *
+     * @returns {boolean} True if the group is touched, false otherwise.
+     */
+    isGroupTouched(groupId) {
+        return this.categories.some((c) => c.groupId === groupId && c.touched);
+    }
+
     isAllowAcceptableAdsFilterEnabled() {
         const { SearchAndSelfPromoFilterId } = this.constants.AntiBannerFiltersId;
         this.isFilterEnabled(SearchAndSelfPromoFilterId);
@@ -307,6 +332,40 @@ class SettingsStore {
         this.isFilterEnabled(UrlTrackingFilterId);
     }
 
+    /**
+     * List of annoyances filters.
+     */
+    @computed
+    get annoyancesFilters() {
+        const annoyancesGroup = this.categories.find((group) => {
+            return group.groupId === AntibannerGroupsId.AnnoyancesFiltersGroupId;
+        });
+        return annoyancesGroup.filters;
+    }
+
+    /**
+     * List of recommended annoyances filters.
+     */
+    @computed
+    get recommendedAnnoyancesFilters() {
+        return this.annoyancesFilters.filter((filter) => {
+            return filter.tags.includes(RECOMMENDED_TAG_ID);
+        });
+    }
+
+    /**
+     * List of AdGuard annoyances filters.
+     */
+    @computed
+    get agAnnoyancesFilters() {
+        return [
+            ...this.recommendedAnnoyancesFilters,
+            this.annoyancesFilters.find((f) => {
+                return f.filterId === AntiBannerFiltersId.AnnoyancesCombinedFilterId;
+            }),
+        ];
+    }
+
     @computed
     get lastUpdateTime() {
         // TODO: lastCheckTime or lastUpdateTime?
@@ -314,15 +373,14 @@ class SettingsStore {
     }
 
     @action
-    async updateGroupSetting(id, enabled) {
-        const recommendedFiltersIds = await messenger.updateGroupStatus(id, enabled);
+    async updateGroupSetting(groupId, enabled) {
+        const recommendedFiltersIds = await messenger.updateGroupStatus(groupId, enabled);
 
         runInAction(() => {
-            const groupId = parseInt(id, 10);
             if (groupId === AntibannerGroupsId.OtherFiltersGroupId
                 && this.isAllowAcceptableAdsFilterEnabled()) {
                 this.allowAcceptableAds = enabled;
-            } else if (groupId === AntibannerGroupsId.PrivacyFilterGroupId) {
+            } else if (groupId === AntibannerGroupsId.PrivacyFiltersGroupId) {
                 if (this.isBlockKnownTrackersFilterEnabled()) {
                     this.blockKnownTrackers = enabled;
                 }
@@ -348,6 +406,37 @@ class SettingsStore {
                 });
             }
         });
+    }
+
+    @action
+    updateGroupStateUI(groupId, enabled) {
+        this.categories.forEach(category => {
+            if (category.groupId === groupId) {
+                if (enabled) {
+                    category.enabled = true;
+                } else {
+                    delete category.enabled;
+                }
+            }
+        });
+    }
+
+    @action
+    updateFilterStateUI(filterId, enabled) {
+        this.filters.forEach(filter => {
+            if (filter.filterId === filterId) {
+                if (enabled) {
+                    filter.enabled = true;
+                } else {
+                    delete filter.enabled;
+                }
+            }
+        });
+    }
+
+    @action
+    setFiltersToGetConsentFor(filters) {
+        this.filtersToGetConsentFor = filters;
     }
 
     @action
@@ -389,8 +478,7 @@ class SettingsStore {
     };
 
     @action
-    async updateFilterSetting(rawFilterId, enabled) {
-        const filterId = Number.parseInt(rawFilterId, 10);
+    async updateFilterSetting(filterId, enabled) {
         this.setFilterEnabledState(filterId, enabled);
         try {
             const groupId = await messenger.updateFilterStatus(filterId, enabled);
@@ -408,6 +496,8 @@ class SettingsStore {
 
                 if (group) {
                     group.enabled = true;
+                    // if any filter in group is enabled, the group is considered as touched
+                    group.touched = true;
                 }
             }
         } catch (e) {
@@ -703,6 +793,53 @@ class SettingsStore {
     get isUpdateFiltersButtonActive() {
         return this.filters.some((filter) => filter.enabled
             && this.isCategoryEnabled(filter.groupId));
+    }
+
+    @action
+    setIsAnnoyancesConsentModalOpen = (value) => {
+        this.isAnnoyancesConsentModalOpen = value;
+    };
+
+    @action
+    setFilterIdSelectedForConsent = (filterId) => {
+        this.filterIdSelectedForConsent = filterId;
+        this.updateFilterStateUI(filterId, true);
+    };
+
+    @action
+    handleFilterConsentCancel = () => {
+        if (this.filterIdSelectedForConsent) {
+            this.updateFilterStateUI(this.filterIdSelectedForConsent, false);
+            this.filterIdSelectedForConsent = null;
+            return;
+        }
+
+        // handle modal for group
+        this.updateGroupStateUI(AntibannerGroupsId.AnnoyancesFiltersGroupId, false);
+    };
+
+    @action
+    handleFilterConsentConfirm = async () => {
+        if (this.filterIdSelectedForConsent) {
+            await this.updateFilterSetting(this.filterIdSelectedForConsent, true);
+            await messenger.setConsentedFilters([this.filterIdSelectedForConsent]);
+            this.filterIdSelectedForConsent = null;
+            return;
+        }
+        // handle consent modal for group
+        await this.updateGroupSetting(AntibannerGroupsId.AnnoyancesFiltersGroupId, true);
+        await messenger.setConsentedFilters(
+            this.recommendedAnnoyancesFilters.map((f) => f.filterId),
+        );
+    };
+
+    @computed
+    get shouldShowFilterPolicy() {
+        if (this.filterIdSelectedForConsent) {
+            return this.agAnnoyancesFilters.some((f) => f.filterId === this.filterIdSelectedForConsent);
+        }
+        // consent modal for group
+        return true;
     }
 }